+++ backend/src/main/java/com/tplaneing/trip_planeing/TripPlaneingApplication.java
