+++ frontend/src/pages/NotesApplet.css
