+++ frontend/src/pages/NotesApplet.jsx
