+++ frontend/src/pages/ReminderApplet.jsx
